refactor(net): split directional handlers
